some anoying warnings

+++ AddConnectionController.m
@@ -244,7 +244,7 @@
      	NSLog(@"doOpen we have a Cancel button");
      	return;
     } else {
-     	NSLog(@"doOpen tvarInt not equal 1 or zero = %3li",tvarNSInteger);
+     	NSLog(@"doOpen tvarInt not equal 1 or zero = %3d",(int)tvarNSInteger);
      	return;
     } // end if
 }
